inventory v4
created an item tag, an object that has it can be picked up (set up colliders and the raw image with icon first). inventory shows up with I, can be nulled with N. only key numbers go down with U. inventory script is still under development but the base idea is completed. raised the first table a little bit so the character cant move above it. the assets folder has the key model with key icon, and a wine icon for testing (you can pick up the wine bottle thats in the middle of the chairs). there's 3 keys around the beginning area for testing.