scaling and hiding ui
scaled the canvases to match the players screen size and the key ui goes away when using the camera or picture template. can be taken off with inventory ui as well. theres a polaroid canvas that uses the first person mode to take pictures. the directory path and the picture inventory is still under development.