screenshot inventory base
changed all the raw image objects to buttons for future functionalities. camera pictures can be seen from the button objects (P opens the picture inventory). the max amount is 8 atm before crashing (array index amount is limiting). the directory path still has to be changed. remember to use timescale next time, currently the script for inventories r pausing the movement and camera but the actual time is not stopped.