added a keypad to lobby's safebox
the safebox under the main painting between the staircases has a functional keypad now. it includes having a canvas thats hidden and only activated by e. can be closed with esc. new scripts such as: keypadsystem (inside keypadcanvas), numbuttonadd (inside number buttons as a component), numdelete is for c button to empty the number field and numOK is for the ok button that checks if the entered number is correct. after that it handles the interactionscript part where the user gets the lobby key (required to open the corridor) to their inventory. also added a few more papers to hold the code numbers. also fixed the cancel issue with them.